pyproject.toml
[project]
name = "btc-mn-copulas"
version = "0.1.0"
description = "BTC Market-Neutral Copula-Based Pairs Trading Bot for Binance Futures"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-binance>=1.0.19",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "statsmodels>=0.14.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
